Move CacheContactExtensionDecorator to builder project
