+++ src/Samples/Builder.Console/BuilderServiceProvider.cs
@@ -17,7 +17,5 @@ namespace Builder.Console
         public void RegisterService(Type serviceType, object instance) => RegisterSingleton(serviceType, instance);
 
         public void RegisterService(Type serviceType, Func<object> instanceFactory) => RegisterSingleton(serviceType, instanceFactory);
-
-        public void RegisterDecorator(Type serviceType, Func<object> instanceFactory) => RegisterDecorator(serviceType, instanceFactory);
     }
 }
