+++ src/Take.Blip.Client/Activation/IServiceContainer.cs
@@ -24,15 +24,5 @@ namespace Take.Blip.Client.Activation
         /// only when a instance of <paramref name="serviceType"/> is requested.
         /// </param>
         void RegisterService(Type serviceType, Func<object> instanceFactory);
-
-        /// <summary>
-        /// Registers a factory for the decorator instance.
-        /// </summary>
-        /// <param name="serviceType">Type of the service.</param>
-        /// <param name="instanceFactory">
-        /// A factory function which will be called
-        /// only when a instance of <paramref name="serviceType"/> is requested.
-        /// </param>
-        void RegisterDecorator(Type serviceType, Func<object> instanceFactory);
     }
 }
